backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
sqlite3
python-multipart==0.0.6
python-dotenv==1.0.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
PyMuPDF==1.23.8
python-docx==1.1.0
nltk==3.8.1
numpy==1.24.4
pandas==2.1.3
pydantic==2.5.0
pydantic-settings==2.1.0
Pillow==10.1.0
requests==2.31.0
aiofiles==0.24.0
Levenshtein==0.25.0
streamlit==1.28.2
plotly==5.17.0
streamlit-option-menu==0.3.6
streamlit-lottie==0.0.5